config: make animus-config-protocol the canonical home for WorkflowConfig types + YAML parser
Move the WorkflowConfig TYPES and the standardized YAML PARSER into
crates/animus-config-protocol so the protocol crate is the canonical
contract; orchestrator-config (kernel) and the animus-config-yaml plugin
become implementations that depend on it.

- New protocol modules: workflow_types, agent_types (parse-input agent
  types + merge_agent_profile, split out of the kernel's 5389-line
  agent_runtime_config.rs), yaml_types, yaml_parser, yaml_diagnostic,
  builtins, env_interp (env-only), parse (compile_yaml_workflow_files +
  merge_yaml_into_config), overlay (generated-overlay readers/writers).
- env_interp drops the pluggable WorkflowSecretResolver / keychain
  machinery and all secret-interpolation passes: it is env-only now.
  ${secret.*} is NOT resolved at config-parse — it survives verbatim
  into the compiled config and is resolved at consume/spawn time.
- Kernel keeps the COMPILER: validation, pack-overlay merge, state-machine
  derivation, disk cache, loading, config_source_client. agent_runtime_config
  re-exports agent_types; workflow_config re-exports the moved types/parser
  so the hundreds of crate::workflow_config::* call sites are unchanged.
- Bootstrap path: workflow config compile / init now compile via the
  protocol library parser (compile_workflow_config_from_library) with no
  config_source plugin required; the daemon runtime load path still
  requires a plugin. Fixes init_e2e + 'workflow config compile'.
- Spawn-time secret resolution: animus-runtime-shared resolves
  ${secret.<name>} in MCP server env values (anywhere in the scalar,
  honoring the $$ escape) from the secrets: declaration + keychain
  snapshot, so secret-backed MCP env still works without parse-time baking.
- Removed obsolete keychain-resolver / secret-resolution-at-parse tests;
  redirected env tests; removed install_keychain_workflow_resolver_for.
