+++ crates/animus-config-protocol/Cargo.toml
@@ -7,7 +7,11 @@ description = "Wire types for the Animus `config_source` plugin role (JSON-RPC 2
 [dependencies]
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
+serde_yaml = "0.9"
 schemars = "1.0"
+anyhow = "1.0"
+sha2 = "0.10"
+protocol = { workspace = true }
 
 [dev-dependencies]
 tempfile = "3"
