+++ crates/animus-config-protocol/src/yaml_diagnostic.rs
