+++ crates/orchestrator-daemon-runtime/src/daemon/run_daemon.rs
@@ -184,11 +184,13 @@ where
     crate::quotas::install_runtime_quota_process_slot_factory();
 
     // v0.5.8 secrets: wire keychain-backed secrets into the plugin spawn
-    // path and the workflow YAML interpolator. Both installs are
-    // first-installer-wins, so tests that pre-install a mock keep theirs.
+    // path. First-installer-wins, so tests that pre-install a mock keep theirs.
+    // v0.6: the workflow-YAML interpolator is env-only — `${secret.*}` is no
+    // longer resolved at config-parse time, so there is no keychain resolver to
+    // install here. Secrets resolve at consume/spawn time via the snapshot
+    // provider above.
     let secrets_project_root = std::path::Path::new(project_root);
     let _ = crate::quotas::install_keychain_secret_provider_for(secrets_project_root);
-    let _ = crate::quotas::install_keychain_workflow_resolver_for(secrets_project_root);
 
     hooks.handle_event(DaemonRunEvent::Startup { project_root: primary_root.clone(), daemon_pid })?;
